fix(assistant): update assistant to pass tests
