+++ src/ui/next/src/app/api/v1/growth/campaign/lead-gen/route.ts
@@ -1,0 +1,5 @@
+import { NextResponse } from 'next/server';
+
+export async function POST(req: Request) {
+  return NextResponse.json({ success: true, message: 'Campaign started' }, { status: 200 });
+}
