+++ src/ui/next/src/e2e/assistant.spec.ts
@@ -3,8 +3,13 @@ import { test, expect } from '@playwright/test';
 
 test.describe('Assistant Page', () => {
   test('navigates to assistant and verifies authentic state', async ({ page }) => {
-    await adminPage({ page }, async ({ page }) => {
-      await page.goto('/assistant');
+    await page.goto('/login');
+    await page.getByPlaceholder('Email or Username').fill('test@example.com');
+    await page.getByPlaceholder('Password').fill('password123');
+    await page.getByRole('button', { name: 'Log In' }).click();
+    await expect(page.getByRole('heading', { name: 'Dashboard' }).first()).toBeVisible();
+
+    await page.goto('/assistant');
 
       // Verify the page shell and layout
       await expect(page.getByTestId('assistant-shell')).toBeVisible();
@@ -27,6 +32,5 @@ test.describe('Assistant Page', () => {
       // Check a panel (e.g., Remote Control)
       await page.getByRole('button', { name: 'Remote Control' }).click();
       await expect(page.getByRole('heading', { name: 'Remote Control' })).toBeVisible();
-    });
   });
 });
