+++ src/ui/next/test-results/.last-run.json
@@ -1,6 +1,4 @@
 {
-  "status": "failed",
-  "failedTests": [
-    "9b8cd1bc48a798eff244-8f813b7d7d2b664d3987"
-  ]
+  "status": "passed",
+  "failedTests": []
 }
